fix #4299 ユーザーが存在するユーザーグループを削除できないようにする (#4329)
- UserGroupsService::delete() で所属ユーザーの存在確認を追加
- ユーザーが所属している場合は BcException をスローして削除を拒否
- API/管理画面コントローラーにエラーハンドリングを追加
- ユニットテストと統合テストを追加

+++ plugins/baser-core/src/Service/UserGroupsService.php
@@ -11,6 +11,7 @@
 
 namespace BaserCore\Service;
 
+use BaserCore\Error\BcException;
 use BaserCore\Model\Entity\UserGroup;
 use BaserCore\Model\Table\UserGroupsTable;
 use BaserCore\Utility\BcContainerTrait;
@@ -172,13 +173,17 @@ class UserGroupsService implements UserGroupsServiceInterface
      *
      * @param int $id
      * @return bool
+     * @throws BcException ユーザーが所属している場合
      * @checked
      * @noTodo
      * @unitTest
      */
     public function delete(int $id): bool
     {
-        $userGroup = $this->UserGroups->get($id);
+        $userGroup = $this->UserGroups->get($id, ['contain' => ['Users']]);
+        if (!empty($userGroup->users)) {
+            throw new BcException(__d('baser_core', 'ユーザーが所属しているユーザーグループは削除できません。'));
+        }
         return $this->UserGroups->delete($userGroup);
     }
